# openapi-live.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "s-curve",
    "description": "",
    "contact": {
      "name": "s-curve team"
    },
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8800"
    },
    {
      "url": "https://rust-service:8800"
    }
  ],
  "paths": {
...
    "/auth/forgot-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Request password reset",
        "description": "Generates a password reset token and returns it. In production, this should send an email.",
        "operationId": "forgot_password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              },
              "example": {
                "email": "user@example.com"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Reset token generated (for dev; in prod would send email)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                },
                "example": {
                  "message": "Operation successful"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": []
      },
      "post": {
        "description": "Generates a password reset token and returns it. In production, this should send an email.",
        "operationId": "forgot_password",
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "email": "user@example.com"
              },
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "message": "Operation successful"
                },
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "Reset token generated (for dev; in prod would send email)"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [],
        "summary": "Request password reset",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login",
        "description": "Login Handles `POST` requests for `/auth/login`. Does not require authentication.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "example": {
                "email": "user@example.com",
                "password": "password123"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                },
                "example": {
                  "token": "eyJhbGciOiJIUzI1Ni...",
                  "user": {
                    "created_at": "2025-01-15T10:00:00Z",
                    "email": "ada@example.com",
                    "id": "aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa",
                    "name": "Ada Lovelace",
                    "provider": "local",
                    "updated_at": "2025-01-15T10:00:00Z"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "security": []
      },
      "post": {
        "description": "Login Handles `POST` requests for `/auth/login`. Does not require authentication.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "email": "user@example.com",
                "password": "password123"
              },
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "token": "eyJhbGciOiJIUzI1Ni...",
                  "user": {
                    "created_at": "2025-01-15T10:00:00Z",
                    "email": "ada@example.com",
                    "id": "aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa",
                    "name": "Ada Lovelace",
                    "provider": "local",
                    "updated_at": "2025-01-15T10:00:00Z"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            },
            "description": "Login successful"
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "security": [],
        "summary": "Login",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Logout",
        "description": "Logout Handles `POST` requests for `/auth/logout`. Requires bearer authentication.",
        "operationId": "logout",
        "responses": {
          "200": {
            "description": "Logout acknowledged"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "description": "Logout Handles `POST` requests for `/auth/logout`. Requires bearer authentication.",
        "operationId": "logout",
        "responses": {
          "200": {
            "description": "Logout acknowledged"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Logout",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Me",
        "description": "Me Handles `GET` requests for `/auth/me`. Requires bearer authentication.",
        "operationId": "me",
        "responses": {
          "200": {
            "description": "Current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                },
                "example": {
                  "created_at": "2025-01-15T10:00:00Z",
                  "email": "ada@example.com",
                  "id": "aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa",
                  "name": "Ada Lovelace",
                  "provider": "local",
                  "updated_at": "2025-01-15T10:00:00Z"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "get": {
        "description": "Me Handles `GET` requests for `/auth/me`. Requires bearer authentication.",
        "operationId": "me",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "created_at": "2025-01-15T10:00:00Z",
                  "email": "ada@example.com",
                  "id": "aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa",
                  "name": "Ada Lovelace",
                  "provider": "local",
                  "updated_at": "2025-01-15T10:00:00Z"
                },
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "description": "Current user"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Me",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register",
        "description": "Register Handles `POST` requests for `/auth/register`. Does not require authentication.",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              },
              "example": {
                "email": "test@example.com",
                "name": "Test User",
                "password": "password123"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "User registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                },
                "example": {
                  "token": "eyJhbGciOiJIUzI1Ni...",
                  "user": {
                    "created_at": "2025-01-15T10:00:00Z",
                    "email": "ada@example.com",
                    "id": "aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa",
                    "name": "Ada Lovelace",
                    "provider": "local",
                    "updated_at": "2025-01-15T10:00:00Z"
                  }
                }
              }
            }
          },
          "409": {
            "description": "Email already in use"
          }
        },
        "security": []
      },
      "post": {
        "description": "Register Handles `POST` requests for `/auth/register`. Does not require authentication.",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "email": "test@example.com",
                "name": "Test User",
                "password": "password123"
              },
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "example": {
                  "token": "eyJhbGciOiJIUzI1Ni...",
                  "user": {
                    "created_at": "2025-01-15T10:00:00Z",
                    "email": "ada@example.com",
                    "id": "aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa",
                    "name": "Ada Lovelace",
                    "provider": "local",
                    "updated_at": "2025-01-15T10:00:00Z"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            },
            "description": "User registered"
          },
          "409": {
            "description": "Email already in use"
          }
        },
        "security": [],
        "summary": "Register",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/reset-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Reset password with token",
        "description": "Validates the reset token and sets a new password.",
        "operationId": "reset_password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              },
              "example": {
                "new_password": "NewSecurePassword456!",
                "token": "raw-token-from-email"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Password reset successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                },
                "example": {
                  "message": "Operation successful"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired token"
          }
        },
        "security": []
      },
      "post": {
        "description": "Validates the reset token and sets a new password.",
        "operationId": "reset_password",
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "new_password": "NewSecurePassword456!",
                "token": "raw-token-from-email"
              },
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "message": "Operation successful"
                },
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "Password reset successful"
          },
          "400": {
            "description": "Invalid or expired token"
          }
        },
        "security": [],
        "summary": "Reset password with token",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthResponse": {
        "type": "object",
        "required": [
          "token",
          "user"
        ],
        "properties": {
          "token": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
...
      "ForgotPasswordRequest": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "example": "ada@example.com"
          },
          "password": {
            "type": "string",
            "example": "S3cureP@ssw0rd"
          }
        }
      },
      "MessageResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "RegisterRequest": {
        "type": "object",
        "required": [
          "name",
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "example": "ada@example.com"
          },
          "name": {
            "type": "string",
            "example": "Ada Lovelace"
          },
          "password": {
            "type": "string",
            "example": "S3cureP@ssw0rd"
          }
        }
      },
      "ResetPasswordRequest": {
        "type": "object",
        "required": [
          "token",
          "new_password"
        ],
        "properties": {
          "new_password": {
            "type": "string",
            "example": "NewSecureP@ss456"
          },
          "token": {
            "type": "string",
            "example": "abc123token"
          }
        }
      },
...
      "User": {
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "provider",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "provider_id": {
            "type": "string",
            "nullable": true
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}